% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admit_reason_description_count.R
\name{admit_reason_description_count}
\alias{admit_reason_description_count}
\title{Show each unique Admit_Reason_Description Details}
\usage{
admit_reason_description_count(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
Show each unique Admit_Reason_Description Details
}
