\name{admit_source}
\alias{admit_source}
\docType{data}
\title{
Admission Source
}
\description{
List of admission sources
}
\usage{data("admit_source")}
\format{
  A data frame with 9 observations on the following 9 variables.
  \describe{
    \item{\samp{Concept Code}}{a numeric vector}
    \item{\samp{Concept Name}}{a character vector}
    \item{\samp{Preferred Concept Name}}{a character vector}
    \item{\samp{Preferred Alternate Code}}{a logical vector}
    \item{\samp{Code System OID}}{a character vector}
    \item{\samp{Code System Name}}{a character vector}
    \item{\samp{Code System Code}}{a character vector}
    \item{\samp{Code System Version}}{a character vector}
    \item{\samp{HL7 Table 0396 Code}}{a character vector}
  }
}
\details{
Needs more documentation
}

\examples{
data(admit_source)
## maybe str(admit_source) ; plot(admit_source) ...
}
\keyword{datasets}
