% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admit_source_invalid.R
\name{admit_source_invalid}
\alias{admit_source_invalid}
\title{Getting Invalid Examples and Summaries for Admit_Source}
\usage{
admit_source_invalid(data)
}
\arguments{
\item{data}{The raw data from BioSense on which you will do the invalid admit source checks.}
}
\value{
A list of two data frames: examples and summary for invalid Admit_Source.
}
\description{
Getting Invalid Examples and Summaries for Admit_Source
}
\examples{
\dontrun{
  admit_source_invalid(df)
}
}
