% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_group_perc.R
\name{age_group_perc}
\alias{age_group_perc}
\title{Get variable C_Patient_Age_Years Frequency}
\usage{
age_group_perc(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
Get variable C_Patient_Age_Years Frequency
}
