\name{age_units}
\alias{age_units}
\docType{data}
\title{
Age Units
}
\description{
Contains list of age unit options
}
\usage{data("age_units")}
\format{
  A data frame with 5 observations on the following 9 variables.
  \describe{
    \item{\samp{Concept Code}}{a character vector}
    \item{\samp{Concept Name}}{a character vector}
    \item{\samp{Preferred Concept Name}}{a character vector}
    \item{\samp{Preferred Alternate Code}}{a character vector}
    \item{\samp{Code System OID}}{a character vector}
    \item{\samp{Code System Name}}{a character vector}
    \item{\samp{Code System Code}}{a character vector}
    \item{\samp{Code System Version}}{a numeric vector}
    \item{\samp{HL7 Table 0396 Code}}{a character vector}
  }
}
\details{
Needs more documentations
}
\examples{
data(age_units)
## maybe str(age_units) ; plot(age_units) ...
}
\keyword{datasets}
