% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_visit_length.R
\name{avg_visit_length}
\alias{avg_visit_length}
\title{compute the average length of each visit}
\usage{
avg_visit_length(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
compute the average length of each visit
}
