% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_visit_per_day.R
\name{avg_visit_per_day}
\alias{avg_visit_per_day}
\title{Compute the average number of visits per day for each facility}
\usage{
avg_visit_per_day(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a numeric vector
}
\description{
Compute the average number of visits per day for each facility
}
