% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_report_admin.R
\name{batch_all_production_admin}
\alias{batch_all_production_admin}
\title{Batch Job To Create NSSP BioSense Platform Data Quality Summary Reports for All Kansas Facilities in Production for the Hospital Admins}
\usage{
batch_all_production_admin(
  facility_spreadsheet,
  table,
  mft,
  username,
  password,
  start,
  end,
  directory,
  email = F,
  sender,
  email_password,
  personname = NA,
  title = NA,
  phone = NA
)
}
\arguments{
\item{facility_spreadsheet}{location of the facility spreadsheet xlsx}

\item{table}{The table that you want to retrieve the data from, as a string.}

\item{mft}{The MFT (master facilities table) from where the facility name will be retrieved, as a string.}

\item{username}{Your BioSense username, as a string. This is the same username you may use to log into RStudio or Adminer.}

\item{password}{Your BioSense password, as a string. This is the same password you may use to log into RStudio or Adminer.}

\item{start}{The start date time that you wish to begin pulling data from, as a string.}

\item{end}{The end data time that you wish to stop pulling data from, as a string.}

\item{directory}{The directory where you would like to write the reports to (i.e., "~/Documents/MyReports"), as a string.}

\item{email}{Default False. If True, then the function will atempt to send out a form}

\item{sender}{Email address of sender. Make sure it's kdhe.KS.gov}

\item{email_password}{Your Email Password}

\item{personname}{Your Name to be used in your email text}

\item{title}{Your job title to be used in your email text}

\item{phone}{Your phone number to be used in your email text}
}
\value{
Report table stored at directory location. If email=TRUE, then a email will be sent. A table with facility, receiver and conformation of email being sent.
}
\description{
This function iteratively performs `write_facilty_report`  function for all Kansas Facilities in production targeted toward admins. It will generate summary report for all specified facility.
The summary workbook shows percents and counts of nulls and invalids, Additionally it generates a timeliness
report and creates a table.
}
\examples{
library(emayili)
library(ggplot2)
library(readxl)
library(keyring)

## store passwords for essence
key_set(service = "essence")
## store passwords for email
key_set(service = "email")
batch_all_production_admin(facility_spreadsheet="Facilities Spreadsheet.xlsx", username="bzhang02", password=key_get("essence"),
 table="KS_PR_Processed", mft="KS_MFT",  start="2020-05-01 00:00:00", 
 end="2020-05-30 23:59:59", directory="~",
 email =TRUE,sender="bo.zhang@kdhe.ks.gov",
email_password=key_get("email"),personname='Bo Zhang',title='intern',phone='630-457-8915')
}
