% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_info.R
\name{batch_info}
\alias{batch_info}
\title{compute the number of batches per day and the average time between batches for each facility}
\usage{
batch_info(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
compute the number of batches per day and the average time between batches for each facility
}
