\name{blood_pressure_units}
\alias{blood_pressure_units}
\docType{data}
\title{
Blood Pressure Units
}
\description{
Contains a list of blood pressure units
}
\usage{data("blood_pressure_units")}
\format{
  A data frame with 1 observations on the following 9 variables.
  \describe{
    \item{\samp{Concept.Code}}{a character vector}
    \item{\samp{Concept.Name}}{a character vector}
    \item{\samp{Preferred Concept Name}}{a character vector}
    \item{\samp{Preferred Alternate Code}}{a logical vector}
    \item{\samp{Code System OID}}{a character vector}
    \item{\samp{Code System Name}}{a character vector}
    \item{\samp{Code System Code}}{a character vector}
    \item{\samp{Code System Version}}{a numeric vector}
    \item{\samp{HL7 Table 0396 Code}}{a character vector}
  }
}
\details{
Needs more documentation
}
\examples{
data(blood_pressure_units)
## maybe str(blood_pressure_units) ; plot(blood_pressure_units) ...
}
\keyword{datasets}
