% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chief_complaint_text_count.R
\name{chief_complaint_text_count}
\alias{chief_complaint_text_count}
\title{Chief Complaint Text Count}
\usage{
chief_complaint_text_count(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
Chief Complaint Text Count
}
