% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_impression_count.R
\name{clinical_impression_count}
\alias{clinical_impression_count}
\title{Clinical Impression Count}
\usage{
clinical_impression_count(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
Clinical Impression Count
}
