\name{country}
\alias{country}
\docType{data}
\title{
Global Country List
}
\description{
This is a look-up table for countries around the world.
}
\usage{data("country")}
\format{
  A data frame with 246 observations on the following 9 variables.
  \describe{
    \item{\samp{Concept Code}}{a character vector}
    \item{\samp{Concept Name}}{a character vector}
    \item{\samp{Preferred Concept Name}}{a character vector}
    \item{\samp{Preferred Alternate Code}}{a character vector}
    \item{\samp{Code System OID}}{a character vector}
    \item{\samp{Code System Name}}{a character vector}
    \item{\samp{Code System Code}}{a character vector}
    \item{\samp{Code System Version}}{a character vector}
    \item{\samp{HL7 Table 0396 Code}}{a character vector}
  }
}
\details{
Needs more documentation
}
\examples{
data(country)
## maybe str(country) ; plot(country) ...
}
\keyword{datasets}
