\name{county}
\alias{county}
\docType{data}
\title{
List of US Counties
}
\description{
Countains a list of US Counties
}
\usage{data("county")}
\format{
  A data frame with 3281 observations on the following 3 variables.
  \describe{
    \item{\samp{Concept Code}}{a character vector}
    \item{\samp{Concept Name}}{a character vector}
    \item{\samp{Preferred Concept Name}}{a character vector}
  }
}
\details{
Needs more documentation
}
\examples{
data(county)
## maybe str(county) ; plot(county) ...
}
\keyword{datasets}
