% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_invalid.R
\name{date_time_invalid}
\alias{date_time_invalid}
\title{Check the validity of the format of all _data_time related variables}
\usage{
date_time_invalid(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a list
}
\description{
Check the validity of the format of all _data_time related variables
}
