% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis_code_invalid.R
\name{diagnosis_code_invalid}
\alias{diagnosis_code_invalid}
\title{Check the validity of diagnosis code}
\usage{
diagnosis_code_invalid(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a list
}
\description{
Check the validity of diagnosis code
}
