% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis_code_perc.R
\name{diagnosis_code_perc}
\alias{diagnosis_code_perc}
\title{Get Diagnosis_code Frequency Percent}
\usage{
diagnosis_code_perc(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
Get Diagnosis_code Frequency Percent
}
