\name{diagnosis_type}
\alias{diagnosis_type}
\docType{data}
\title{
List of Diagnostis Types
}
\description{
Look-up table of diagnosis types
}
\usage{data("diagnosis_type")}
\format{
  A data frame with 3 observations on the following 9 variables.
  \describe{
    \item{\samp{Concept Code}}{a character vector}
    \item{\samp{Concept Name}}{a character vector}
    \item{\samp{Preferred Concept Name}}{a character vector}
    \item{\samp{Preferred Alternate Code}}{a logical vector}
    \item{\samp{Code System OID}}{a character vector}
    \item{\samp{Code System Name}}{a character vector}
    \item{\samp{Code System Code}}{a character vector}
    \item{\samp{Code System Version}}{a character vector}
    \item{\samp{HL7 Table 0396 Code}}{a character vector}
  }
}
\details{
Needs more documentation
}
\examples{
data(diagnosis_type)
## maybe str(diagnosis_type) ; plot(diagnosis_type) ...
}
\keyword{datasets}
