% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis_type_perc.R
\name{diagnosis_type_perc}
\alias{diagnosis_type_perc}
\title{Get Diagnosis Type Frequency}
\usage{
diagnosis_type_perc(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
Get Diagnosis Type Frequency
}
