\name{discharge_disposition}
\alias{discharge_disposition}
\docType{data}
\title{
List of Discharge Dispositions
}
\description{
Look-up table for discharge dispositions
}
\usage{data("discharge_disposition")}
\format{
  A data frame with 109 observations on the following 9 variables.
  \describe{
    \item{\samp{Concept Code}}{a character vector}
    \item{\samp{Concept Name}}{a character vector}
    \item{\samp{Preferred Concept Name}}{a character vector}
    \item{\samp{Preferred Alternate Code}}{a logical vector}
    \item{\samp{Code System OID}}{a character vector}
    \item{\samp{Code System Name}}{a character vector}
    \item{\samp{Code System Code}}{a character vector}
    \item{\samp{Code System Version}}{a character vector}
    \item{\samp{HL7 Table 0396 Code}}{a character vector}
  }
}
\details{
Needs more documentation}
\examples{
data(discharge_disposition)
## maybe str(discharge_disposition) ; plot(discharge_disposition) ...
}
\keyword{datasets}
