% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge_disposition_perc.R
\name{discharge_disposition_perc}
\alias{discharge_disposition_perc}
\title{Get Frequency of discharge disposition}
\usage{
discharge_disposition_perc(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
Get Frequency of discharge disposition
}
