% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/early_lag.R
\name{early_lag}
\alias{early_lag}
\title{Compute the time difference for the earliest Recorded_Date_Time}
\usage{
early_lag(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
Compute the time difference for the earliest Recorded_Date_Time
}
