\name{ethnicity}
\alias{ethnicity}
\docType{data}
\title{
List of Ethnicities
}
\description{
Look-up table for ethnicities
}
\usage{data("ethnicity")}
\format{
  A data frame with 2 observations on the following 9 variables.
  \describe{
    \item{\samp{Concept.Code}}{a character vector}
    \item{\samp{Concept.Name}}{a character vector}
    \item{\samp{Preferred Concept Name}}{a character vector}
    \item{\samp{Preferred Alternate Code}}{a logical vector}
    \item{\samp{Code System OID}}{a character vector}
    \item{\samp{Code System Name}}{a character vector}
    \item{\samp{Code System Code}}{a character vector}
    \item{\samp{Code System Version}}{a numeric vector}
    \item{\samp{HL7 Table 0396 Code}}{a character vector}
  }
}
\details{
Needs more documentation
}
\examples{
data(ethnicity)
## maybe str(ethnicity) ; plot(ethnicity) ...
}
\keyword{datasets}
