% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethnicity_code_perc.R
\name{ethnicity_code_perc}
\alias{ethnicity_code_perc}
\title{Get Frequency of Ethnicity Codes}
\usage{
ethnicity_code_perc(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
Get Frequency of Ethnicity Codes
}
