% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethnicity_description_perc.R
\name{ethnicity_description_perc}
\alias{ethnicity_description_perc}
\title{Ethnicity_Description Percent}
\usage{
ethnicity_description_perc(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
Ethnicity_Description Percent
}
