\name{facility_type}
\alias{facility_type}
\docType{data}
\title{
Facility Types
}
\description{
Look-up table for types of facility
}
\usage{data("facility_type")}
\format{
  A data frame with 5 observations on the following 9 variables.
  \describe{
    \item{\samp{Concept.Code}}{a character vector}
    \item{\samp{Concept.Name}}{a character vector}
    \item{\samp{Preferred Concept Name}}{a character vector}
    \item{\samp{Preferred Alternate Code}}{a logical vector}
    \item{\samp{Code System OID}}{a character vector}
    \item{\samp{Code System Name}}{a character vector}
    \item{\samp{Code System Code}}{a character vector}
    \item{\samp{Code System Version}}{a numeric vector}
    \item{\samp{HL7 Table 0396 Code}}{a character vector}
  }
}
\details{
Needs more documentation
}
\examples{
data(facility_type)
## maybe str(facility_type) ; plot(facility_type) ...
}
\keyword{datasets}
