% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facility_type_code_perc.R
\name{facility_type_code_perc}
\alias{facility_type_code_perc}
\title{Facility Type Code Percent}
\usage{
facility_type_code_perc(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
Facility Type Code Percent
}
