\name{height_units}
\alias{height_units}
\docType{data}
\title{
Height Units
}
\description{
Look-up table for height units
}
\usage{data("height_units")}
\format{
  A data frame with 4 observations on the following 9 variables.
  \describe{
    \item{\samp{Concept.Code}}{a character vector}
    \item{\samp{Concept.Name}}{a character vector}
    \item{\samp{Preferred Concept Name}}{a character vector}
    \item{\samp{Preferred Alternate Code}}{a logical vector}
    \item{\samp{Code System OID}}{a character vector}
    \item{\samp{Code System Name}}{a character vector}
    \item{\samp{Code System Code}}{a character vector}
    \item{\samp{Code System Version}}{a numeric vector}
    \item{\samp{HL7 Table 0396 Code}}{a character vector}
  }
}
\details{
Needs more documentation
}
\examples{
data(height_units)
## maybe str(height_units) ; plot(height_units) ...
}
\keyword{datasets}
