\name{hl7_values}
\alias{hl7_values}
\docType{data}
\title{
HL7 Values
}
\description{
Look-up table of HL7 Values
}
\usage{data("hl7_values")}
\format{
  A data frame with 89 observations on the following 2 variables.
  \describe{
    \item{\code{HL7}}{a character vector}
    \item{\code{Field}}{a character vector}
  }
}
\details{
Needs more documentation
}
\examples{
data(hl7_values)
## maybe str(hl7_values) ; plot(hl7_values) ...
}
\keyword{datasets}
