% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insurance_company_id_perc.R
\name{insurance_company_id_perc}
\alias{insurance_company_id_perc}
\title{Insurance Company ID Percent}
\usage{
insurance_company_id_perc(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
Insurance Company ID Percent
}
