% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_by_trigger.R
\name{lag_by_trigger}
\alias{lag_by_trigger}
\title{Raise Lag by Trigger Event}
\usage{
lag_by_trigger(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
Raise Lag by Trigger Event
}
