% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_chief_complaint.R
\name{lag_chief_complaint}
\alias{lag_chief_complaint}
\title{Calculate lag between Chief Complaints and other Variables
this function will compute the lag between message recorded and visit time, between message sent and message recorded time, arrival
and message sent time, arrival and visit time, for the earliest recorded non NA C_Chief_Complaints}
\usage{
lag_chief_complaint(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
Calculate lag between Chief Complaints and other Variables
this function will compute the lag between message recorded and visit time, between message sent and message recorded time, arrival
and message sent time, arrival and visit time, for the earliest recorded non NA C_Chief_Complaints
}
