% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_graph.R
\name{lag_graph}
\alias{lag_graph}
\title{Lag Graph}
\usage{
lag_graph(data, start, end)
}
\arguments{
\item{data}{a data frame}

\item{start}{a date vector}

\item{end}{a date vector}
}
\value{
a graph
}
\description{
Lag Graph
}
