% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_message_per_batch.R
\name{mean_message_per_batch}
\alias{mean_message_per_batch}
\title{compute the average number of messages}
\usage{
mean_message_per_batch(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
compute the average number of messages
}
