\name{patient_class}
\alias{patient_class}
\docType{data}
\title{
Patient Classes
}
\description{
Look-up table of classes of patients
}
\usage{data("patient_class")}
\format{
  A data frame with 8 observations on the following 9 variables.
  \describe{
    \item{\samp{Concept.Code}}{a character vector}
    \item{\samp{Concept.Name}}{a character vector}
    \item{\samp{Preferred Concept Name}}{a character vector}
    \item{\samp{Preferred Alternate Code}}{a logical vector}
    \item{\samp{Code System OID}}{a character vector}
    \item{\samp{Code System Name}}{a character vector}
    \item{\samp{Code System Code}}{a character vector}
    \item{\samp{Code System Version}}{a character vector}
    \item{\samp{HL7 Table 0396 Code}}{a character vector}
  }
}
\details{
Needs more documentation
}
\examples{
data(patient_class)
## maybe str(patient_class) ; plot(patient_class) ...
}
\keyword{datasets}
