% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_class_perc.R
\name{patient_class_perc}
\alias{patient_class_perc}
\title{Frequency of occurrence in data - what is actually occurring in the data}
\usage{
patient_class_perc(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
Frequency of occurrence in data - what is actually occurring in the data
}
