\name{pulse_oximetry_units}
\alias{pulse_oximetry_units}
\docType{data}
\title{
Pulse Oximetry Units
}
\description{
Look-up table of pulse oximetry units
}
\usage{data("pulse_oximetry_units")}
\format{
  A data frame with 1 observations on the following 9 variables.
  \describe{
    \item{\samp{Concept.Code}}{a character vector}
    \item{\samp{Concept.Name}}{a character vector}
    \item{\samp{Preferred Concept Name}}{a character vector}
    \item{\samp{Preferred Alternate Code}}{a logical vector}
    \item{\samp{Code System OID}}{a character vector}
    \item{\samp{Code System Name}}{a character vector}
    \item{\samp{Code System Code}}{a character vector}
    \item{\samp{Code System Version}}{a numeric vector}
    \item{\samp{HL7 Table 0396 Code}}{a character vector}
  }
}
\details{
Needs more documentation
}
\examples{
data(pulse_oximetry_units)
## maybe str(pulse_oximetry_units) ; plot(pulse_oximetry_units) ...
}
\keyword{datasets}
