\name{race}
\alias{race}
\docType{data}
\title{
Race Categories
}
\description{
Look-up table of race categories
}
\usage{data("race")}
\format{
  A data frame with 6 observations on the following 9 variables.
  \describe{
    \item{\samp{Concept Code}}{a character vector}
    \item{\samp{Concept Name}}{a character vector}
    \item{\samp{Preferred Concept Name}}{a character vector}
    \item{\samp{Preferred Alternate Code}}{a logical vector}
    \item{\samp{Code System OID}}{a character vector}
    \item{\samp{Code System Name}}{a character vector}
    \item{\samp{Code System Code}}{a character vector}
    \item{\samp{Code System Version}}{a numeric vector}
    \item{\samp{HL7 Table 0396 Code}}{a character vector}
  }
}
\details{
Needs more documentation
}
\examples{
data(race)
## maybe str(race) ; plot(race) ...
}
\keyword{datasets}
