% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_code_perc.R
\name{race_code_perc}
\alias{race_code_perc}
\title{Frequency of occurrence in data - what is actually occurring in the data}
\usage{
race_code_perc(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
Frequency of occurrence in data - what is actually occurring in the data
}
