% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/race_description_perc.R
\name{race_description_perc}
\alias{race_description_perc}
\title{Calculate frequencies for Race_Description Variable}
\usage{
race_description_perc(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
Calculate frequencies for Race_Description Variable
}
