% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_facility_table_excel.R
\name{repeated_facility_excel}
\alias{repeated_facility_excel}
\title{Write NSSP BioSense Platform Data Quality Summary Reports for Multiple Facility, Use excel as input}
\usage{
repeated_facility_excel(
  input,
  contact,
  table = NA,
  mft = NA,
  username = NA,
  password = NA,
  start = NA,
  end = NA,
  directory = NA,
  field = NA,
  exclude = NA,
  optional = T,
  email = NA,
  sender = NA,
  email_password = NA,
  personname = NA,
  title = NA,
  phone = NA
)
}
\arguments{
\item{input}{location of input.xlsx file.}

\item{contact}{A datatable of 2 colomns, facility is the facility number, receiver is the receive email address}

\item{table}{The table that you want to retrieve the data from, as a string.}

\item{mft}{The MFT (master facilities table) from where the facility name will be retrieved, as a string.}

\item{username}{Your BioSense username, as a string. This is the same username you may use to log into RStudio or Adminer.}

\item{password}{Your BioSense password, as a string. This is the same password you may use to log into RStudio or Adminer.}

\item{start}{The start date time that you wish to begin pulling data from, as a string.}

\item{end}{The end data time that you wish to stop pulling data from, as a string.}

\item{directory}{The directory where you would like to write the reports to (i.e., "~/Documents/MyReports"), as a string.}

\item{field}{Default NA. Can add a string with delimiter of ':'. Only fields that countain those words will be included in the final report.}

\item{exclude}{Default NA. Can add a string with delimiter of ':'. Exclude fields with certain keywords in the final report.}

\item{optional}{Default True. If False then remove all optional fields}

\item{email}{Default False. If True, then the function will atempt to send out a form}

\item{sender}{Email address of sender. Make sure it's kdhe.KS.gov}

\item{email_password}{Your Email Password}

\item{personname}{Your Name to be used in your email text}

\item{title}{Your job title to be used in your email text}

\item{phone}{Your phone number to be used in your email text}
}
\value{
Report table stored at directory location. If email=TRUE, then a email will be sent. A table with facility, receiver and conformation of email being sent.
}
\description{
This performs `write_facilty_report`  function for all . It will generate summary report for all specified facility.
This function uses excel generated information but can be override by additionall inputs
The summary workbook shows percents and counts of nulls and invalids, Additionally it generates a timeliness
report and creates a table. The program can send out a report to designated email address
}
\examples{
library(emayili)
library(ggplot2)
library(readxl)
repeated_facility_excel("Input.xlsx", contact = contact)
##you can override fields from the input.xlsx
repeated_facility_excel("Input.xlsx", contact = contact,field='age',email=F)
}
