\name{smoking_status}
\alias{smoking_status}
\docType{data}
\title{
Smoking Status
}
\description{
Look-up table of smoking status values
}
\usage{data("smoking_status")}
\format{
  A data frame with 9 observations on the following 9 variables.
  \describe{
    \item{\samp{Concept Code}}{a numeric vector}
    \item{\samp{Concept Name}}{a character vector}
    \item{\samp{Preferred Concept Name}}{a character vector}
    \item{\samp{Preferred Alternate Code}}{a character vector}
    \item{\samp{Code System OID}}{a character vector}
    \item{\samp{Code System Name}}{a character vector}
    \item{\samp{Code System Code}}{a character vector}
    \item{\samp{Code System Version}}{a numeric vector}
    \item{\samp{HL7 Table 0396 Code}}{a character vector}
  }
}
\details{
Needs more documentation
}
\examples{
data(smoking_status)
## maybe str(smoking_status) ; plot(smoking_status) ...
}
\keyword{datasets}
