% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking_status_code_perc.R
\name{smoking_status_code_perc}
\alias{smoking_status_code_perc}
\title{Calculate Percentages for smoking_status_code}
\usage{
smoking_status_code_perc(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
Calculate Percentages for smoking_status_code
}
