% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking_status_description_perc.R
\name{smoking_status_description_perc}
\alias{smoking_status_description_perc}
\title{Frequency of occurrence in data - what is actually occurring in the data}
\usage{
smoking_status_description_perc(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
Frequency of occurrence in data - what is actually occurring in the data
}
