\name{state}
\alias{state}
\docType{data}
\title{
States and US Territories
}
\description{
Look-up table of US states and territories
}
\usage{data("state")}
\format{
  A data frame with 69 observations on the following 9 variables.
  \describe{
    \item{\samp{Concept.Code}}{a character vector}
    \item{\samp{Concept.Name}}{a character vector}
    \item{\samp{Preferred Concept Name}}{a character vector}
    \item{\samp{Preferred Alternate Code}}{a character vector}
    \item{\samp{Code System OID}}{a character vector}
    \item{\samp{Code System Name}}{a character vector}
    \item{\samp{Code System Code}}{a character vector}
    \item{\samp{Code System Version}}{a numeric vector}
    \item{\samp{HL7 Table 0396 Code}}{a character vector}
  }
}
\details{
Needs more documentation
}
\examples{
data(state)
## maybe str(state) ; plot(state) ...
}
\keyword{datasets}
