% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_perc.R
\name{state_perc}
\alias{state_perc}
\title{Frequency of Occurrence in the Data}
\usage{
state_perc(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
Frequency of Occurrence in the Data
}
