% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triage_notes_count.R
\name{triage_notes_count}
\alias{triage_notes_count}
\title{Show each unique Triage_Notes for each visit}
\usage{
triage_notes_count(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
Show each unique Triage_Notes for each visit
}
