% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigger_event_perc.R
\name{trigger_event_perc}
\alias{trigger_event_perc}
\title{Frequency of occurrence in data}
\usage{
trigger_event_perc(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
Frequency of occurrence in data
}
