% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va_lag.R
\name{va_lag}
\alias{va_lag}
\title{Computes Average Time Difference between Item Pairs}
\usage{
va_lag(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
Computes Average Time Difference between Item Pairs
}
\examples{
\dontrun{
  dfsum <- va_lag(df)
}
}
