% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_facility_bo_write_in.R
\name{write_facility_bo_write_in}
\alias{write_facility_bo_write_in}
\title{Write Facility Write-In}
\usage{
write_facility_bo_write_in()
}
\value{
a data frame produced by write_facility_report
}
\description{
Write Facility Write-In
}
