% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_facility_bo_excel.R
\name{write_facility_excel}
\alias{write_facility_excel}
\title{Write NSSP BioSense Platform Data Quality Summary Reports for one Facility, Use excel as input}
\usage{
write_facility_excel(input)
}
\arguments{
\item{input}{location of input.xlsx file.}
}
\description{
This performs `write_facilty_report`  function for all . It will generate summary report for all specified facility.
This function uses excel generated information 
The summary workbook shows percents and counts of nulls and invalids, Additionally it generates a timeliness
report and creates a table. The program can send out a report to designated email address
}
\examples{
library(emayili)
library(ggplot2)
library(readxl)
write_facility_excel("Input.xlsx")
}
