% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_facility_it.R
\name{write_facility_it}
\alias{write_facility_it}
\title{Write NSSP BioSense Platform Data Quality Summary Reports for One Facility}
\usage{
write_facility_it(
  username,
  password,
  table,
  mft,
  start,
  end,
  facility,
  directory = "",
  field = NA,
  exclude = NA,
  optional = TRUE,
  email = FALSE,
  sender = NA,
  receiver = NA,
  email_password = NA,
  personname = NA,
  title = NA,
  phone = NA,
  message = NA,
  facility_type = NULL
)
}
\arguments{
\item{username}{Your BioSense username, as a string. This is the same username you may use to log into RStudio or Adminer.}

\item{password}{Your BioSense password, as a string. This is the same password you may use to log into RStudio or Adminer.}

\item{table}{The table that you want to retrieve the data from, as a string.}

\item{mft}{The MFT (master facilities table) from where the facility name will be retrieved, as a string.}

\item{start}{The start date time that you wish to begin pulling data from, as a string.}

\item{end}{The end data time that you wish to stop pulling data from, as a string.}

\item{facility}{The C_Biosense_Facility_ID for the facility that you wish to generate and write the report for.}

\item{directory}{The directory where you would like to write the reports to (i.e., "~/Documents/MyReports"), as a string.}

\item{field}{Default NA. Can add a string with delimiter of ':'. Only fields that countain those words will be included in the final report.}

\item{exclude}{Default NA. Can add a string with delimiter of ':'. Exclude fields with certain keywords in the final report.}

\item{optional}{Default True. If False then remove all optional fields}

\item{email}{Default False. If True, then the function will atempt to send out a form}

\item{sender}{Email address of sender. Make sure it's kdhe.KS.gov}

\item{receiver}{Email address of receiver.}

\item{email_password}{Your Email Password}

\item{personname}{Your Name to be used in your email text}

\item{title}{Your job title to be used in your email text}

\item{phone}{Your phone number to be used in your email text}

\item{message}{The email message to be sent. Allows for composition of costume messages.}

\item{facility_type}{a character vector indicating the facility type}
}
\value{
A report table stored at directory location. If email=TRUE, then a email will be sent, along with a confirmation of email being sent.
}
\description{
This function is a `write_reports` function. It will generate summary report for one specified facility.
  The summary workbook shows percents and counts of nulls and invalids, Additionally it generates a timeliness
  report and creates a table. The program can send out a report to designated email address
}
