% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_facility_bo.R
\name{write_facility_report}
\alias{write_facility_report}
\title{Write NSSP BioSense Platform Data Quality Summary Reports for One Facility}
\usage{
write_facility_report(
  username,
  password,
  table,
  mft,
  start,
  end,
  facility,
  directory = "",
  field = NA,
  exclude = NA,
  optional = TRUE,
  email = FALSE,
  sender = NA,
  receiver = NA,
  email_password = NA,
  personname = NA,
  title = NA,
  phone = NA,
  message = NA
)
}
\arguments{
\item{username}{Your BioSense username, as a string.}

\item{password}{Your BioSense password, as a string.}

\item{table}{The table that you want to retrieve the data from, as a string.}

\item{mft}{The MFT (master facilities table)}

\item{start}{The start date time that you wish to begin pulling data from, as a string.}

\item{end}{The end data time that you wish to stop pulling data from, as a string.}

\item{facility}{The C_Biosense_Facility_ID for the facility that you wish to generate}

\item{directory}{The directory where you would like to write the reports to}

\item{field}{Default NA. Can add a string with delimiter of}

\item{exclude}{Default NA. Can add a string with delimiter of}

\item{optional}{Default True. If False then remove all optional fields}

\item{email}{Default False. If True, then the function will atempt to send out a form}

\item{sender}{Email address of sender. Make sure it's kdhe.KS.gov}

\item{receiver}{Email address of receiver.}

\item{email_password}{Your Email Password}

\item{personname}{Your Name to be used in your email text}

\item{title}{Your job title to be used in your email text}

\item{phone}{Your phone number to be used in your email text}

\item{message}{The email message to be sent. Allows for composition of costume messages.}
}
\value{
A report table stored at directory location.
}
\description{
Write NSSP BioSense Platform Data Quality Summary Reports for One Facility
}
