% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_facility_table_excel.R
\name{repeated_facility_excel}
\alias{repeated_facility_excel}
\title{Write NSSP BioSense Platform Data Quality Summary Reports for Multiple Facility, Use excel as input}
\usage{
repeated_facility_excel(input, contact)
}
\arguments{
\item{input}{location of input.xlsx file.}

\item{contact}{A datatable of 2 colomns, facility is the facility number, receiver is the receive email address}
}
\value{
Report table stored at directory location. If email=TRUE, then a email will be sent. A table with facility, receiver and conformation of email being sent.
}
\description{
This performs `write_facilty_report`  function for all . It will generate summary report for all specified facility.
This function uses excel generated information 
The summary workbook shows percents and counts of nulls and invalids, Additionally it generates a timeliness
report and creates a table. The program can send out a report to designated email address
}
\examples{
library(emayili)
library(ggplot2)
library(readxl)
repeated_facility_excel("Input.xlsx", contact = contact)
}
