% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_facility_bo_excel.R
\name{write_facility_excel}
\alias{write_facility_excel}
\title{Write NSSP BioSense Platform Data Quality Summary Reports for one Facility, Use excel as input}
\usage{
write_facility_excel(
  input,
  facility = NA,
  receiver = NA,
  table = NA,
  mft = NA,
  username = NA,
  password = NA,
  start = NA,
  end = NA,
  directory = NA,
  field = NA,
  exclude = NA,
  optional = T,
  email = NA,
  sender = NA,
  email_password = NA,
  personname = NA,
  title = NA,
  phone = NA,
  message = NA
)
}
\arguments{
\item{input}{location of input.xlsx file.}

\item{facility}{The C_Biosense_Facility_ID for the facility that you wish to generate and write the report for.}

\item{receiver}{Email address of receiver.}

\item{table}{The table that you want to retrieve the data from, as a string.}

\item{mft}{The MFT (master facilities table) from where the facility name will be retrieved, as a string.}

\item{username}{Your BioSense username, as a string. This is the same username you may use to log into RStudio or Adminer.}

\item{password}{Your BioSense password, as a string. This is the same password you may use to log into RStudio or Adminer.}

\item{start}{The start date time that you wish to begin pulling data from, as a string.}

\item{end}{The end data time that you wish to stop pulling data from, as a string.}

\item{directory}{The directory where you would like to write the reports to (i.e., "~/Documents/MyReports"), as a string.}

\item{field}{Default NA. Can add a string with delimiter of ':'. Only fields that countain those words will be included in the final report.}

\item{exclude}{Default NA. Can add a string with delimiter of ':'. Exclude fields with certain keywords in the final report.}

\item{optional}{Default True. If False then remove all optional fields}

\item{email}{Default False. If True, then the function will atempt to send out a form}

\item{sender}{Email address of sender. Make sure it's kdhe.KS.gov}

\item{email_password}{Your Email Password}

\item{personname}{Your Name to be used in your email text}

\item{title}{Your job title to be used in your email text}

\item{phone}{Your phone number to be used in your email text
@param message The email message to be sent. Allows for composition of costume messages.}
}
\description{
This performs `write_facilty_report`  function for all . It will generate summary report for all specified facility.
This function uses excel generated information although one can override the information from the input.xlsx
The summary workbook shows percents and counts of nulls and invalids, Additionally it generates a timeliness
report and creates a table. The program can send out a report to designated email address
}
\examples{
library(emayili)
library(ggplot2)
library(readxl)
write_facility_excel("Input.xlsx")
}
